requirements.txt
SpeechRecognition==3.8.1
python-telegram-bot==12.4.2